requirements.txt
pycrypto==2.6.1
pyDes==2.0.1
